--- a/ff/util.go
+++ b/ff/util.go
@@ -1,0 +1,5 @@
+package ff
+
+func NewElement() *Element {
+	return &Element{}
+}