@ed255 2020-03-04
The code looks good to me!  Now the only remaining thing to merge this is resolve any doubt about licensing.
- All our code is published under GPLv3 and the copyright is held by 0kims (thus giving 0kims the possibility of distributing under a different license).
- Goff is Apache 2, which allows commercial distribution without disclosing the source code with the only requirements of:
  - Showing the Apache 2 copyright notice with the distribution
  - Stating the list of changes made to the code

So I think Apache 2 would cause any problem in either the GPLv3 licensed distribution of our libraries nor a possible commercial distribution under a private license (as long as the two requirements of the Apache 2 for Goff are met).